Major commit, cmake can now export targets correctly BUT
Only DataFrame is exported correctly for some reason, but the same
files and config are used for DMScu, which seems the export the
correct files but external projectes can't find it for some reason.

I am commiting to work of this and to save the current work as this is
the correct direction.